use the puzzle shell layout
