+++ frontend/angular.json
@@ -31,7 +31,8 @@
               "src/assets"
             ],
             "styles": [
-              "src/styles.scss"
+              "src/styles.scss",
+              "node_modules/@puzzleitc/puzzle-shell/dist/bundle.js"
             ],
             "scripts": []
           },
