+++ frontend/src/app/app.component.html
@@ -1,9 +1,44 @@
-<div class="content" role="main">
+<pzsh-container>
 
-    <app-header></app-header>
+  <pzsh-topbar>
+    <svg class="headerAndFooter" xmlns="http://www.w3.org/2000/svg" width="214" height="32" fill="none">
+      <path fill="#3FA8E0" d="M107.914 20.016c0-.443-.156-.782-.469-1.016-.312-.24-.875-.49-1.687-.75-.813-.266-1.456-.526-1.93-.781-1.292-.698-1.937-1.638-1.937-2.82 0-.615.171-1.162.515-1.641.349-.485.847-.862 1.492-1.133.651-.27 1.381-.406 2.188-.406.812 0 1.536.148 2.172.445.635.292 1.127.706 1.476 1.242.355.537.532 1.146.532 1.828h-2.344c0-.52-.164-.924-.492-1.21-.328-.292-.789-.438-1.383-.438-.573 0-1.018.122-1.336.367-.318.24-.477.557-.477.953 0 .37.185.68.555.93.375.25.925.484 1.649.703 1.333.401 2.304.898 2.914 1.492.609.594.914 1.334.914 2.219 0 .984-.373 1.758-1.118 2.32-.744.558-1.747.836-3.007.836-.875 0-1.672-.159-2.391-.476-.719-.323-1.268-.763-1.648-1.32-.375-.558-.563-1.204-.563-1.938h2.352c0 1.255.75 1.883 2.25 1.883.557 0 .992-.112 1.304-.336.313-.23.469-.547.469-.953zm12.875-.805c-.088 1.224-.542 2.187-1.359 2.89-.813.704-1.886 1.055-3.219 1.055-1.458 0-2.607-.49-3.445-1.468-.834-.985-1.25-2.334-1.25-4.047v-.696c0-1.093.192-2.057.578-2.89.385-.834.935-1.472 1.648-1.914.719-.448 1.552-.672 2.5-.672 1.313 0 2.37.351 3.172 1.054.802.704 1.266 1.69 1.391 2.961h-2.344c-.057-.734-.263-1.265-.617-1.593-.349-.334-.883-.5-1.602-.5-.781 0-1.367.28-1.758.843-.385.558-.583 1.425-.593 2.602v.86c0 1.228.185 2.127.554 2.695.375.567.964.851 1.766.851.724 0 1.263-.164 1.617-.492.36-.333.565-.846.617-1.54h2.344zm5.75-.375h-1.867V23h-2.344V11.625h4.227c1.343 0 2.38.3 3.109.898.729.6 1.094 1.446 1.094 2.54 0 .775-.169 1.424-.508 1.945-.333.515-.841.927-1.523 1.234l2.461 4.649V23h-2.516l-2.133-4.164zm-1.867-1.898h1.89c.589 0 1.045-.149 1.368-.446.323-.302.484-.716.484-1.242 0-.537-.154-.958-.461-1.266-.302-.307-.768-.46-1.398-.46h-1.883v3.414zm16.484-5.313v7.492c0 1.245-.39 2.23-1.172 2.953-.776.724-1.838 1.086-3.187 1.086-1.328 0-2.383-.351-3.164-1.054-.781-.704-1.18-1.67-1.195-2.899v-7.578h2.343v7.508c0 .745.177 1.289.531 1.633.36.338.855.507 1.485.507 1.318 0 1.987-.692 2.008-2.078v-7.57h2.351zm4.985 0l2.921 8.25 2.907-8.25h3.078V23h-2.352v-3.11l.235-5.367L149.859 23h-1.609l-3.062-8.469.234 5.36V23h-2.344V11.625h3.063zm17.281 7.367V23h-2.344V11.625h4.438c.854 0 1.604.156 2.25.469.651.312 1.151.758 1.5 1.336.349.573.523 1.226.523 1.96 0 1.115-.383
+                     1.995-1.148 2.641-.761.64-1.815.961-3.164.961h-2.055zm0-1.898h2.094c.619 0 1.091-.146 1.414-.438.328-.291.492-.708.492-1.25 0-.557-.164-1.008-.492-1.351-.328-.344-.782-.521-1.36-.532h-2.148v3.57zm17.328.476c0 1.12-.198 2.102-.594 2.946-.396.843-.963 1.494-1.703 1.953-.734.458-1.578.687-2.531.687-.943 0-1.784-.226-2.524-.68-.739-.453-1.312-1.098-1.718-1.937-.407-.844-.612-1.812-.618-2.906v-.563c0-1.12.201-2.104.602-2.953.406-.854.977-1.508
+                     1.711-1.96.74-.46 1.583-.688 2.531-.688.948 0 1.789.229 2.524.687.739.453 1.31 1.107 1.711 1.961.406.85.609 1.83.609 2.945v.508zm-2.375-.515c0-1.193-.214-2.1-.641-2.72-.427-.619-1.036-.929-1.828-.929-.786 0-1.393.307-1.82.922-.427.61-.643 1.505-.648 2.688v.554c0 1.162.213 2.063.64 2.703.427.641 1.042.961 1.844.961.786 0 1.39-.307 1.812-.922.422-.62.636-1.52.641-2.703v-.554zm7.641 1.383l-1.219 1.312V23h-2.344V11.625h2.344v5.156l1.031-1.414 2.899-3.742h2.882l-4.039 5.055 4.157 6.32h-2.789l-2.922-4.563zm13.437-.368h-4.5v3.047h5.281V23h-7.625V11.625h7.61v1.898h-5.266v2.711h4.5v1.836zm6.367.766h-1.867V23h-2.344V11.625h4.227c1.344 0 2.38.3 3.109.898.729.6 1.094 1.446 1.094 2.54 0 .775-.169 1.424-.508 1.945-.333.515-.841.927-1.523 1.234l2.461 4.649V23h-2.516l-2.133-4.164zm-1.867-1.898h1.891c.588 0 1.044-.149 1.367-.446.323-.302.484-.716.484-1.242 0-.537-.153-.958-.461-1.266-.302-.307-.768-.46-1.398-.46h-1.883v3.414z"/>
+      <path fill="#fff" d="M27.45 4.686c-6.28-6.248-16.461-6.248-22.74 0-6.28 6.249-6.28 16.38 0 22.628 6.279 6.248 16.46 6.248 22.74 0 3.015-3 4.71-7.07 4.71-11.314a15.96 15.96 0 00-4.71-11.314zm-1.24 5.527C25.686 11.77 24 16.442 24 16.442c-.564 1.523-3.863 1.51-8.342 1.523 0 0 2.894-.407 3.73-2.015a38.073 38.073 0 001.822-4.864c-4.455 0-4.853 1.553-4.853 1.553s-2.814 7.836-3.82 10.727c-7.804 0-6.557-3.835-6.557-3.835L9.884 8.635h7.905l-.527 1.57s.643-1.57 5.002-1.57c.002 0 4.471.019 3.946 1.578zM41.45 19.15v4.023h-2.746V11.439h4.787a4.907 4.907 0 013.329 1.067 3.519 3.519 0 011.231 2.8 3.505 3.505 0 01-1.231 2.793 4.938 4.938 0 01-3.329 1.059l-2.042-.008zm0-2.095h2.041c.501.035.994-.142 1.356-.488.316-.34.484-.79.465-1.253a1.802 1.802 0 00-.465-1.281 1.763 1.763 0 00-1.356-.5H41.45v3.522zM58.88 11.437v7.689a3.805 3.805 0 01-1.34 3.112 5.347 5.347 0 01-3.51 1.103 5.403 5.403 0 01-3.54-1.103 3.8 3.8 0 01-1.34-3.112v-7.69h2.747v7.69a2.086 2.086 0 00.559 1.6c.439.37 1.005.556 1.58.52a2.185 2.185 0 001.558-.516 2.111 2.111 0 00.547-1.6v-7.693h2.74zM63.33 21.077h5.37v2.096h-8.74v-1.226l5.273-8.414h-5.297v-2.096h8.691v1.177l-5.298 8.463zM72.927 21.077h5.37v2.096h-8.74v-1.226l5.274-8.414h-5.297v-2.096h8.69v1.177l-5.297 8.463zM82.07
+                     21.077h5.079v2.096h-7.823V11.439h2.747l-.003 9.638zM95.471 18.151H90.85v2.926h5.45v2.096h-8.2V11.439h8.213v2.096H90.85v2.523h4.625l-.004 2.093z"/>
+    </svg>
 
+    <span class="spacer"></span>
+
+    <img class="header-link" *ngIf="isCurrentlyOnPlayground()" (click)="changeSpectator()" src="assets/visibility_white_48dp.svg" height="30" width="30" alt="toggle spectator mode">
+
+    <img class="header-link" routerLink="info" src="assets/info_white_48dp.svg" height="30" width="30" alt="show info page">
+
+  </pzsh-topbar>
+
+  <main>
     <router-outlet></router-outlet>
+  </main>
 
-    <app-footer></app-footer>
+  <pzsh-footer>
+    <div slot="start">
+      Version 0.1
+    </div>
 
-</div>
+    <div slot="center">
+      <app-footer></app-footer>
+    </div>
+
+    <div slot="end">
+      <p>Testlinks:
+        <a routerLink="/">/</a>,&nbsp;&nbsp;&nbsp;&nbsp;
+        <a routerLink="/onboarding">/onborarding</a>,&nbsp;&nbsp;&nbsp;&nbsp;
+        <a routerLink="/playground">/playground</a>,&nbsp;&nbsp;&nbsp;&nbsp;
+        <a routerLink="/info">/info</a>,&nbsp;&nbsp;&nbsp;&nbsp;
+        <a routerLink="/error">/error</a>
+      </p>
+    </div>
+  </pzsh-footer>
+
+</pzsh-container>
