+++ frontend/src/app/app.component.scss
@@ -1,8 +1,12 @@
-.content {
-    display: flex;
-    margin: 82px auto 32px;
-    padding: 0 16px;
-    max-width: 960px;
-    flex-direction: column;
-    align-items: center;
+main {
+  padding: var(--pzsh-spacer);
+}
+
+.spacer {
+  width: 100%;
+}
+
+.header-link {
+  cursor: pointer;
+  margin-left: var(--pzsh-spacer);
 }
