+++ frontend/src/app/app.module.ts
@@ -1,9 +1,8 @@
-import { NgModule } from '@angular/core';
+import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { HeaderComponent } from './components/header/header.component';
 import { FooterComponent } from './components/footer/footer.component';
 import { OnboardingComponent } from './components/onboarding/onboarding.component';
 import { ErrorComponent } from './components/error/error.component';
@@ -19,7 +18,6 @@ import { MobileEstimationComponent } from './components/mobile-estimation/mobile
 @NgModule({
     declarations: [
         AppComponent,
-        HeaderComponent,
         FooterComponent,
         OnboardingComponent,
         ErrorComponent,
@@ -38,7 +36,8 @@ import { MobileEstimationComponent } from './components/mobile-estimation/mobile
     ],
     providers: [],
     exports: [CommonModule],
-    bootstrap: [AppComponent]
+    bootstrap: [AppComponent],
+    schemas: [ CUSTOM_ELEMENTS_SCHEMA ]
 })
 export class AppModule {
 }
