--- frontend/src/app/components/footer/footer.component.scss
@@ -1,4 +1,0 @@
-:host {
-    background-color: #1976d2;
-    width: 100%;
-}
