--- frontend/src/app/components/header/header.component.html
@@ -1,11 +1,0 @@
-<div class="containter">
-    <div>
-    <p>Puzzle Scrum Poker</p>
-    </div>
-    <div *ngIf="isCurrentlyOnPlayground()">
-        <img (click)="changeSpectator()" src="assets/visibility_white_48dp.svg" height="30" width="30">
-    </div>
-    <div>
-        <img src="assets/info_white_48dp.svg" height="30" width="30" routerLink="/info">
-    </div>
-</div>
