--- frontend/src/app/components/header/header.component.scss
@@ -1,11 +1,0 @@
-:host {
-    background-color: #1976d2;
-    width: 100%;
-}
-
-.containter {
-    display: grid;
-    grid-template-columns: auto auto auto;
-    align-content: center;
-    align-items: center;
-}
