+++ frontend/src/app/components/playground/playground.component.html
@@ -5,7 +5,4 @@
 
 <app-desktop-estimation></app-desktop-estimation>
 
-
 <button *ngIf="displayButton" class="middleButton" (click)="changeGameState()">Runde start/stopp</button>
-
-
