+++ frontend/src/index.html
@@ -1,13 +1,13 @@
 <!doctype html>
 <html lang="de">
-<head>
+  <head>
     <meta charset="utf-8">
     <title>Puzzle ScrumPoker</title>
     <base href="/">
     <meta content="width=device-width, initial-scale=1" name="viewport">
     <link href="favicon.ico" rel="icon" type="image/x-icon">
-</head>
-<body>
-<app-root></app-root>
-</body>
+  </head>
+  <body>
+    <app-root></app-root>
+  </body>
 </html>
